Refactor for Gnome 46 (#2)
(Almost) completely refactored code base.
We now instantiate `Panel` directly and deal with the fallout later,
instead of copying half of the gnome shell and imitating the original
thing. This is a much simpler approach, that is IMO more maintainable,
but there might be issues later on with signals and stale references.
Oh, well...

Removed everything that didn't seem relevant in order to keep the code
base light. I chose to ignore Chesterton's fences since the original
code is 10 years old and basically unmaintainable at this point.

---------

Signed-off-by: lazanet <lazanet96@gmail.com>